refactor: rename Research.py to research.py and update import
